refactor(showtime): ShowTime 관련 네이밍 및 API 엔드포인트 정리

+++ src/main/java/com/seatwise/showtime/dto/response/SeatAvailabilityResponse.java
@@ -2,4 +2,4 @@ package com.seatwise.showtime.dto.response;
 
 import com.seatwise.seat.domain.SeatGrade;
 
-public record SeatAvailabilityResponse(SeatGrade grade, int totalSeats, int availableSeats) {}
+public record SeatAvailabilityResponse(SeatGrade grade, Long totalSeats, Long availableSeats) {}
